Implement Waypoints
A Waypoint is a gray object that teleports the Rectangle to Monochrome (specifically the spawn location in it). Since saving is only possible in Monochrome, Waypoints act as a convenient alternative to navigating back through the entire level just to save the progress.

Waypoints also reset the color of the environment so that the feature cannot be exploited.