Improve quests system
- Add onUpdate events.
- Add QuestManager for macro-management and future (de)serialization.
- Streamline quest sequencing with two new lists in QuestData.

Notes on quest sequencing:
NextQuests list in QuestData is for quests that QuestManager will try to make available (not active) after given quest is completed (may fail if prerequisites are not met). Prerequisites list is for those prerequisites.
Use case: sequence of non-mandatory quests. If the first quest is listed in prerequisites of the second and then is not completed by the player, the second quest will never become available.