Implement black square AI
Black squares have 3 states: Idle, Pursuing and Alert, which are changed based on whether the square can see the player (i.e. player is within sight range and nothing is obstructing the view) and on configurable timeout periods. Last known player's position updates once in a configurable time period, if the player is visible.
- In Idle state the square does nothing except detect player proximity. Detecting the player puts the square in the Pursuing state.
- In Pursuing state the square actively propels itself towards the player's last known position. It also has doubled detection range in this state. When the pusruing timeout ends without updates on player position, square enters the Alert state.
- In Alert state the detection range is still doubled, but the square is not actively moving anywhere. If it detects the player in this state, it will move to Pursuing state again. If it doesn't for a set period of time, it becomes Idle and resets to starting position.